+++ config/splat.pspeu.dra.yaml
@@ -53,6 +53,7 @@ segments:
       - [0x66B0, c, dra_psp/66B0]
       - [0x67F0, c, dra_psp/67F0]
       - [0x86A0, c, dra_psp/86A0]
+      - [0xA710, c, dra_psp/A710]
       - [0xAA50, c, dra_psp/AA50]
       - [0xB240, c, dra_psp/B240]
       - [0xBDE0, c, dra_psp/BDE0]
@@ -184,7 +185,7 @@ segments:
       - [0xEE3A0, .rodata, dra/624DC]
       - [0xEE3C0, .rodata, dra_psp/60F8]
       - [0xEE3E8, .rodata, dra_psp/65E0]
-      - [0xEE408, .rodata, dra_psp/86A0]
+      - [0xEE408, .rodata, dra_psp/A710]
       - [0xEE428, .rodata, dra_psp/C0B0]
       - [0xEE4C8, .rodata, dra_psp/11C50]
       - [0xEE4E8, .rodata, dra_psp/13080]
